Moved code to repository
Added 2 versions of the same project to the repository.
Both versions collect data from multiple files, in multiple locations,
and generate a summary of the data collected, and write it to a file.

Both versions are written in Rust, and use the same data files.
Both versions use a timer to measure the time taken to collect and
generate the summary of the data.

Version 1:
    - Does not use threads to collect data from multiple files.

Version 2:
    - Uses 40 threads to collect data from multiple files.